Byld_ds_assignment/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
transformers==4.44.0
torch==2.3.1
numpy==1.26.4
